Move Highlight file to utils and rename to MarkdownHighlighter.
